private key generation is the crashing beast here, with 0x1A9E's method we can store the private key in a 2.1KB Space, it COULD work!

+++ blue-app-iota/src/ternary_store.c
@@ -80,6 +80,6 @@ void ternary_store_set_trit(uint16_t index, trit_t trit, uint8_t out[])
   out[arrayIndex] += bits;
 }
 
-int ternary_store_calculate_array_length(uint16_t amountOfTrits) {
+uint16_t ternary_store_calculate_array_length(uint16_t amountOfTrits) {
   return ceil(amountOfTrits / 4);
 }
