+++ blue-app-iota/src/ternary_store.h
@@ -15,7 +15,7 @@
 #include "../src/vendor/iota/iota_types.h"
 void ternary_store_set_trit(uint16_t index, trit_t trit, uint8_t out[]);
 int8_t ternary_store_get_trit(uint16_t index, uint8_t in[]);
-int ternary_store_calculate_array_length(uint16_t amountOfTrits);
+uint16_t ternary_store_calculate_array_length(uint16_t amountOfTrits);
 uint8_t ternary_store_trit_to_bit(trit_t trit);
 trit_t ternary_store_bit_to_trit(uint8_t bit);
 #endif
